core of emmy tool works
Window is created. Upper panes populate with influence and pop center
info. Clicking on a pop center on the map causes that info to update and
the text window to populate with all your emissaries and their chances
to perform a one or two step transition on that pop center.