Added in-range indicator to emissary tool
Added a column to the text in the emissary tool text box which indicates
if the emissary on that row is in range of the targeted population
center. It says 'no' if the distance is greater than 10, 'yes', if it is
less than 10, and 'HERE!' if the emissary in in that population center
already.