Toggling Markers
Added item to the Map menu allowing you to toggle the exploration
markers on and off
Added item to the Map menu allowing you to toggle showing groups on and
off
The showAllGroups button still works as before.
Bringing up the dialog to add exploration markers will toggle them on.